ai_agents/main_v2.py
"""
Sales Call Analyzer V2 - Proper LangGraph Agentic System
Main execution script with LangSmith integration
"""

import os
import sys
from pathlib import Path

# Set LangSmith environment variables BEFORE importing langchain
os.environ["LANGSMITH_TRACING"] = os.getenv("LANGSMITH_TRACING", "true")
os.environ["LANGSMITH_ENDPOINT"] = os.getenv("LANGSMITH_ENDPOINT", "https://api.smith.langchain.com")
os.environ["LANGSMITH_API_KEY"] = os.getenv("LANGSMITH_API_KEY", "")
os.environ["LANGSMITH_PROJECT"] = os.getenv("LANGSMITH_PROJECT", "Sales_call_analysis_agent")

# Load environment variables
from dotenv import load_dotenv
load_dotenv()

print(f"""
================================================================================
SALES CALL ANALYZER V2 - LANGGRAPH AGENTIC SYSTEM
================================================================================
LangSmith Tracing: {os.getenv('LANGSMITH_TRACING')}
LangSmith Project: {os.getenv('LANGSMITH_PROJECT')}
Model: gemini-2.0-flash-exp
================================================================================
""")

print("[INFO] System starting - this is a proper LangGraph agentic framework")
print("[INFO] Features: Structured output, LangSmith tracing, proper state management")
print("[INFO] Please wait while the system initializes...")

# TODO: Import and run the v2 system
# This is a placeholder - full implementation follows

print("\n[ERROR] V2 system is being built - please use this as the entry point")
print("[INFO] The complete rebuild is in progress to fix all identified issues")
print("\nIssues being fixed:")
print("  1. Gemini schema compliance (using structured output)")
print("  2. None value bugs (proper error handling)")
print("  3. True LangGraph architecture (nodes, state, tools)")
print("  4. LangSmith integration (tracing enabled)")
print("  5. Proper logging (flush handlers)")
print("\nEstimated completion: 30-45 minutes for full rebuild")

