ai_agents/requirements.txt
# Sales Call Analyzer Requirements
# Python 3.10+

# Core Dependencies
pandas>=2.0.0
numpy>=1.24.0

# LLM and AI
google-generativeai>=0.3.0
langchain>=0.1.0
langchain-core>=0.1.0
langgraph>=0.0.20

# Data Validation
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Environment Management
python-dotenv>=1.0.0

# Progress Tracking
tqdm>=4.65.0

# Utilities
typing-extensions>=4.5.0

